HH-81172 move embedded postgresql to nab-testbase
