+++ nab-testbase/src/main/java/ru/hh/nab/testbase/NabTestConfig.java
@@ -1,16 +1,17 @@
-package ru.hh.nab.starter;
+package ru.hh.nab.testbase;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import ru.hh.nab.common.properties.FileSettings;
+import ru.hh.nab.starter.NabCommonConfig;
 
 import java.util.Properties;
 
 @Configuration
 @Import({NabCommonConfig.class})
 public class NabTestConfig {
-  static final String TEST_SERVICE_NAME = "testService";
+  public static final String TEST_SERVICE_NAME = "testService";
   static final String TEST_DATACENTER = "testDC";
 
   @Bean
